backend/pyproject.toml
[project]
name = "trimodal-xai-backend"
version = "1.0.0"
description = "FastAPI backend for tri-modal medicinal leaf classification with XAI"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "tensorflow>=2.15.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scikit-image>=0.21.0",
    "rembg>=2.0.50",
    "onnxruntime>=1.16.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
]

